Move text::Buffer and related types to their own module
